requirements.txt
mysql-connector-python==8.0.33
tkcalendar==1.6.1
Pillow==10.2.0
requests==2.31.0
cryptography==42.0.2 